+++ java/stable/java/org/futo/inputmethod/latin/CrashLoggingApplication.kt
@@ -13,6 +13,7 @@ import org.acra.config.mailSender
 import org.acra.data.StringFormat
 import org.acra.ktx.initAcra
 import androidx.core.content.edit
+import org.futo.inputmethod.latin.uix.isDirectBootUnlocked
 
 class CrashLoggingApplication : Application() /*, Configuration.Provider*/ {
     //override val workManagerConfiguration: Configuration
@@ -21,12 +22,22 @@ class CrashLoggingApplication : Application() /*, Configuration.Provider*/ {
     override fun attachBaseContext(base: Context?) {
         super.attachBaseContext(base)
 
-        if (getSharedPreferences("migrate", MODE_PRIVATE).getBoolean("wiped_work", false) == false) {
-            deleteDatabase("androidx.work.workdb")
-            getSharedPreferences("androidx.work.util.preferences", MODE_PRIVATE)
-                .edit { clear() }
-            getSharedPreferences("migrate", MODE_PRIVATE)
-                .edit { putBoolean("wiped_work", true) }
+        if(isDirectBootUnlocked) {
+            try {
+                if (getSharedPreferences("migrate", MODE_PRIVATE).getBoolean(
+                        "wiped_work",
+                        false
+                    ) == false
+                ) {
+                    deleteDatabase("androidx.work.workdb")
+                    getSharedPreferences("androidx.work.util.preferences", MODE_PRIVATE)
+                        .edit { clear() }
+                    getSharedPreferences("migrate", MODE_PRIVATE)
+                        .edit { putBoolean("wiped_work", true) }
+                }
+            } catch(e: Exception) {
+                e.printStackTrace()
+            }
         }
 
         if(BuildConfig.DEBUG) return
